refactor(Util): Add Util.java for basic common methods.
