+++ src/main/java/org/opentripplanner/middleware/OtpRequestGenerator.java
@@ -1,14 +1,10 @@
 package org.opentripplanner.middleware;
 
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
 import java.util.Map;
 
 import static org.opentripplanner.middleware.Const.*;
 
 public class OtpRequestGenerator {
-
-    final static int NUM_ITINERARIES = 3;
 
     public static void main(String[] args) {
         // Just print out a list of OTP requests from the places in the const file.
@@ -44,7 +40,7 @@ public class OtpRequestGenerator {
 
                     for (int k = 0; k < modeParams.size(); k++) {
                         Map.Entry<String, String> m = modeParams.get(k);
-                        String uiUrl = makeOtpRequestUrl(OTP_UI_URL, e1, e2, m.getValue());
+                        String uiUrl = Util.makeOtpRequestUrl(OTP_UI_URL, e1, e2, m.getValue());
                         System.out.println("<a target='_blank' href='" + uiUrl + "'>" + m.getKey() + "</a> ");
                     };
 
@@ -57,22 +53,5 @@ public class OtpRequestGenerator {
         System.out.println("</table>");
         System.out.println("<p>Total combinations: " + count + "</p>");
         System.out.println("</body><html>");
-    }
-
-    public static String getPlaceParam(Map.Entry<String, Double[]> loc) {
-        Double[] arr = loc.getValue();
-        double lat = arr[1];
-        double lon = arr[0];
-
-        return URLEncoder.encode(loc.getKey(), StandardCharsets.UTF_8) + "::" + lat + "," + lon;
-    }
-
-    public static String makeOtpRequestUrl(String baseUrl, Map.Entry<String, Double[]> loc1, Map.Entry<String, Double[]> loc2, String modeParam) {
-        return baseUrl
-                + "?fromPlace=" + getPlaceParam(loc1)
-                + "&toPlace=" + getPlaceParam(loc2)
-                + "&mode=" + modeParam
-                + "&showIntermediateStops=true&maxWalkDistance=1609&optimize=QUICK&walkSpeed=1.34&ignoreRealtimeUpdates=true"
-                + "&numItineraries=" + NUM_ITINERARIES;
     }
 }
